backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
uvicorn[standard]==0.27.0
pydantic==2.5.3
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
bcrypt==5.0.0
cryptography==46.0.4
ecdsa==0.19.1
passlib==1.7.4
pyasn1==0.6.2
python-jose==3.5.0
rsa==4.9.1
